Revert "classi per il cliente"
