--- Kelnero/OrdersCustomerTableViewController.swift
@@ -1,131 +1,0 @@
-//
-//  OrdersCustomerTableViewController.swift
-//  Kelnero
-//
-//  Created by Monaco Simone on 18/07/18.
-//  Copyright © 2018 Team 6_5. All rights reserved.
-//
-
-import UIKit
-
-class OrdersCustomerTableViewController: UITableViewController {
-    
-    var dish = [Ordine] ()
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        dish = [
-            Ordine(quantity: 3, name: "Cioccolata"),
-            Ordine(quantity: 2, name: "Cotoletta"),
-            Ordine(quantity: 5, name: "Polpette al sugo")
-        ]
-
-        // Uncomment the following line to preserve selection between presentations
-        // self.clearsSelectionOnViewWillAppear = false
-
-        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        self.navigationItem.rightBarButtonItem = self.editButtonItem
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        
-        tableView.reloadData()
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return self.dish.count
-    }
-
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "customerCell", for: indexPath) as! OrdersCustomerCell
-
-        // Configure the cell...
-        let custOrder = self.dish[indexPath.row]
-        cell.nameLabel.text = custOrder.name
-        cell.quantityLabel.text = "Quantity: \(custOrder.quantity)"
-        
-        return cell
-    }
-    
-
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
-
-    
-    // Override to support editing the table view.
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            dish.remove(at: indexPath.row)
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        }
-
-    }
-        
-        
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-        let srcItem = dish[fromIndexPath.row]
-
-        dish.remove(at: fromIndexPath.row)
-        dish.insert(srcItem, at: to.row)
-    }
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        
-        switch segue.identifier {
-        case "showOrder"?:
-            break
-        default:
-            preconditionFailure("Invalid segue identifier.")
-        }
-        
-        if segue.identifier == "showOrder" {
-            
-            // Figure out which row was just tapped
-            //            if let row = tableView.indexPathForSelectedRow?.row {
-            
-            // Get the item associated with this row and pass it along
-            //                let item = itemStore.allItems[row]
-            //                let detailViewController = segue.destination as! DetailViewController
-            //                detailViewController.item = item
-            //                detailViewController.imageStore = imageStore
-            //            }
-        }
-    }
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-    }
-
