--- Kelnero/Ordine.swift
@@ -1,19 +1,0 @@
-//
-//  Ordine.swift
-//  Kelnero
-//
-//  Created by Monaco Simone on 18/07/18.
-//  Copyright © 2018 Team 6_5. All rights reserved.
-//
-
-import Foundation
-
-class Ordine{
-    var quantity: Int
-    var name: String
-    
-    init(quantity: Int, name: String){
-        self.quantity = quantity
-        self.name = name
-    }
-}
